backend/requirements.txt
fastapi
uvicorn
yfinance
nsepython
pandas
google-generativeai
python-dotenv
pydantic
requests
